+++ src/Main.java
@@ -2,68 +2,108 @@ import java.util.Arrays;
 import java.util.Random;
 import java.util.Scanner;
 
+
 public class Main {
-    public static void main(String[] args) {
 
+    public static int ScanNum() {
         Scanner scan = new Scanner(System.in);
+        return scan.nextInt();
+    }
 
+    public static int ScanBoardSize() {
         System.out.print("게임 보드의 크기 입력 (5~15) : ");
         int size;
-        size = scan.nextInt();
+        size = ScanNum();
+        final int MinBoardSize = 5 ;
+        final int MaxBoardSize = 15 ;
 
-        while (size<5 || size>15){
+        while (size < MinBoardSize || size > MaxBoardSize) {
             System.out.println("크기 입력 오류");
             System.out.print("게임 보드의 크기 입력 (5~15) : ");
-            size = scan.nextInt();
+            size = ScanNum();
         }
 
-        String[][] arr = new String[size][size];
+        return size;
+    }
 
+    public static int ScanMineNum(int size) {
         System.out.print("지뢰 개수 입력 : ");
-        int x_num = scan.nextInt();
+        int mine_num = ScanNum();
+        final int MinMineNum = size * size / 10 ;
+        final int MaxMineNum = size * size * 2 / 10 ;
 
-        while (x_num<size*size/10 || x_num>size*size*2/10){
+        while (mine_num < MinMineNum || mine_num > MaxMineNum) {
             System.out.println("옳지 않은 개수");
             System.out.print("지뢰 개수 입력 : ");
-            x_num = scan.nextInt();
+            mine_num = ScanNum();
         }
+        return mine_num;
+    }
 
+    public static String[][] MakeBoard(int size, int mine_num) {
+        String[][] arr = new String[size][size];
         for (String[] strings : arr) Arrays.fill(strings, "O");
 
+        return CountMine(RandomMinePlanting(arr,mine_num), size);
+    }
+
+    public static String[][] RandomMinePlanting(String[][] arr, int mine_num) {
         Random random = new Random();
-        for(int i = 0; i < x_num; i++){
-            int x = random.nextInt(0,arr.length);
-            int y = random.nextInt(0,arr.length);
+        for (int i = 0; i < mine_num; i++) {
+            int x = random.nextInt(0, arr.length);
+            int y = random.nextInt(0, arr.length);
             arr[x][y] = "X";
         }
+        return arr;
+    }
 
-        int [] dx = {-1,-1,-1,0,0,1,1,1};
-        int [] dy = {-1,0,1,-1,1,-1,0,1};
-        int cnt =0;
-
-        for(int i = 0 ; i < arr.length ; i++) {
-            for (int j = 0; j < arr[i].length; j++) {
-                if (arr[i][j].equals("O")) {
-                    for (int k = 0; k < 8; k++) {
-                        int next_x = i + dx[k];
-                        int next_y = j + dy[k];
-
-                        if (0 <= next_x && next_x < size && 0 <= next_y && next_y < size) {
-                            if(arr[next_x][next_y].equals("X")){cnt+=1;}
-                        }
-                    }
-                    if (cnt != 0 ){
-                    arr[i][j] = Integer.toString(cnt);
-                    cnt =0;}
-                }
-            }
-        }
-
+    public static void PrintBoard(String[][] arr) {
         for (String[] strings : arr) {
             for (String string : strings) {
                 System.out.print(string);
             }
             System.out.println();
         }
+
+    }
+
+    public static String[][] CountMine(String[][] arr, int size) {
+        for (int i = 0; i < arr.length; i++) {
+            for (int j = 0; j < arr[i].length; j++) {
+                if (!arr[i][j].equals("X")) CountMineOnBlank(arr, size, i, j);
+            }
+        }
+        return arr;
+    }
+
+    public static void CountMineOnBlank(String[][] arr, int size, int x, int y) {
+        int[] dx = {-1, -1, -1, 0, 0, 1, 1, 1};
+        int[] dy = {-1, 0, 1, -1, 1, -1, 0, 1};
+        int cnt = 0;
+
+        for (int k = 0; k < 8; k++) {
+            int next_x = x + dx[k];
+            int next_y = y + dy[k];
+
+            if (CheckInRange(arr,size,next_x,next_y)) cnt += 1;
+        }
+        if(cnt!=0) arr[x][y] = Integer.toString(cnt);
+    }
+
+    public static boolean CheckInRange(String[][] arr ,int size, int next_x, int next_y){
+        boolean CheckMine = false;
+        boolean IsInRange = 0 <= next_x && next_x < size && 0 <= next_y && next_y < size;
+        if (IsInRange) {CheckMine = arr[next_x][next_y].equals("X");}
+        return CheckMine;
+    }
+
+
+    public static void main(String[] args) {
+
+        int size = ScanBoardSize();
+        int mine_num = ScanMineNum(size);
+        String[][] arr = MakeBoard(size, mine_num);
+        PrintBoard(arr);
+
     }
 }
